Refactored compound prediction into new module
