+++ smact/structure_prediction/__init__.py
@@ -1,0 +1,17 @@
+"""Minimalist ionic compound prediction tools for materials design."""
+
+import logging
+
+__author__ = "Alexander Moriarty"
+__credits__ = {
+  "WMD Group",
+  "Imperial College London",
+  "Andrew Jackson",
+  "Dan Davies",
+  "Keith Butler",
+  "Aron Walsh",
+  "Alexander Moriarty"
+}
+__status__ = "Development"
+
+logger = logging.getLogger(__name__)
