+++ smact/tests/test_structure.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python
+"""Test structure prediction module"""
 
 import itertools
 import json
@@ -6,16 +6,17 @@ import logging
 import os
 import pickle
 import unittest
-
 from contextlib import contextmanager
 
 import numpy as np
 import pandas as pd
 import pymatgen
 
+import smact
 from smact import Species
-import smact.builder
-from smact.builder import CationMutator, SmactStructure, StructureDB
+from smact.structure_prediction.database import StructureDB
+from smact.structure_prediction.mutation import CationMutator
+from smact.structure_prediction.structure import SmactStructure
 
 files_dir = os.path.join(os.path.dirname(os.path.realpath(__file__)), "files")
 TEST_STRUCT = os.path.join(files_dir, "test_struct")
@@ -112,7 +113,7 @@ class StructureTest(unittest.TestCase):
             d = json.load(f)
             py_structure = pymatgen.Structure.from_dict(d)
 
-        with ignore_warnings(smact.builder.logger):
+        with ignore_warnings(smact.structure_prediction.logger):
             s1 = SmactStructure.from_py_struct(py_structure)
 
         s2 = SmactStructure.from_file(os.path.join(files_dir, "CaTiO3.txt"))
